refactor: create meal function to throw error

+++ src/storage/meal/createMeal.ts
@@ -2,6 +2,7 @@
 import AsyncStorage from '@react-native-async-storage/async-storage'
 import { MEAL_COLLECTION } from '../storageConfig'
 import { getMeals } from './getMeals'
+import { AppError } from '../../utils/AppError'
 
 export interface MealProps {
   id: string
@@ -15,6 +16,11 @@ export interface MealProps {
 export async function createMeal(meal: MealProps) {
   try {
     const storedMeals = await getMeals()
+    const mealAlreadyExists = storedMeals.includes(meal)
+
+    if (mealAlreadyExists) {
+      throw new AppError('Você já cadastrou uma refeição com esses dados')
+    }
 
     await AsyncStorage.setItem(
       MEAL_COLLECTION,
